requirements.txt
Django==2.2
django-debug-toolbar==1.11
dj-database-url==0.5.0
django-appconf==1.0.3
django-scheduler==0.8.8
django-wkhtmltopdf==3.2.0
django-weasyprint
html5lib==1.0.1
icalendar==4.0.3
Pillow==6.0.0
Pillow==6.0.0
PuLP==1.6.9
psycopg2 == 2.8.2
pycparser==2.19
pyparsing==2.4.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.1
setuptools==40.5.0
six==1.12.0
sqlparse==0.3.0
sortedcontainers==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.3
